# contracts/openapi.v0.1.yaml
openapi: 3.1.0
info:
  title: DSH 工业仿真智能体 工程 API
  version: 0.1.2
  description: |
    dsh-sim R0 工程能力真相入口（定义书 §工程API：业务操作 / §工程API：审查、节点与上传）。
    MCP 是本 API 的受限投影。长操作返回 202 与资源 ID；请求返回不是计算结束。
    所有产生副作用的 POST 必须携带 Idempotency-Key 头（主体+项目+action+key 唯一；
    同 key 同摘要返回原对象，同 key 异摘要 → 409）。
    错误对象统一为 Error{code,message,retryable,trace_id,details}。
    本文件中的示例值均为占位：假哈希一律写作 "sha256:<64hex>" 形式，不含真实工程参数。
servers:
  - url: http://127.0.0.1:8600/api/v1
...
paths:
...
  /verifications:
    get:
      tags: [verifications]
      operationId: listVerifications
      summary: 按 task_id 或 run_id 读取复算记录
      x-extension: read-projection
      parameters:
        - name: task_id
          in: query
          required: false
          schema: { type: string }
        - name: run_id
          in: query
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Verification 列表
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Verification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /verifications/recheck:
    post:
      tags: [verifications]
      operationId: recheckVerification
      summary: 独立复算（不照抄报告）：从原始 artifact 重算派生指标并产生新 Verification 记录
      x-extension: read-projection
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [run_id]
              properties:
                run_id: { type: string }
      responses:
        '201':
          description: 新 Verification 记录
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Verification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
components:
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
      description: 幂等键；主体+项目+action+key 唯一。同 key 同请求摘要返回原对象；同 key 异摘要 → 409。
...
  responses:
    Unauthorized:
      description: 401 未认证
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: 403 越权
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: 409 摘要/修订/幂等冲突
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unprocessable:
      description: 422 工程输入不成立
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ServiceUnavailable:
      description: 503 暂不可用
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      additionalProperties: false
      required: [code, message, retryable, trace_id, details]
      properties:
        code:
          type: string
          description: 字符串枚举错误码
          example: "<ERROR_CODE>"
        message: { type: string }
        retryable:
          type: boolean
          description: 客户端只能对 retryable 与已定义状态采取有界重试
        trace_id: { type: string, format: uuid }
        details: { type: object }
    Sha256:
      type: string
      description: SHA-256 摘要（64 位小写十六进制）；占位写作 sha256:<64hex>。哈希仅由服务计算。
      example: "sha256:<64hex>"
...
    NumericalState:
      enum: [NOT_CHECKED, PASS, FAIL, INSUFFICIENT]
...
    Verification:
      type: object
      required: [verification_id, run_id, rule_set_sha256, conclusion, findings, source_artifact_ids, created_at]
      properties:
        verification_id: { type: string }
        run_id: { type: string }
        attempt_id: { type: string, nullable: true }
        rule_set_sha256: { $ref: '#/components/schemas/Sha256' }
        check_inputs: { type: object, description: 检查输入摘要 }
        conclusion: { $ref: '#/components/schemas/NumericalState' }
        findings:
          type: array
          items: { type: object }
          description: 缺值/NaN/单位错误/来源不明均生成明确 finding
        source_artifact_ids:
          type: array
          items: { type: string }
        created_at: { type: string, format: date-time }
...